requirements.txt
ipykernel
numpy==1.24.1
matplotlib
ultralytics==8.3.12
from_root
flask
Pillow
pyyaml
flask_limiter
-e .